### 2019-11-08 #1879 @whedon
[PRE REVIEW]: PyCHAM: CHemistry with Aerosol Microphysics in Python
**Submitting author:** @simonom (<a href="http://orcid.org/0000-0001-7197-0651">Simon O&#39;Meara</a>)
**Repository:** <a href="https://github.com/simonom/PyCHAM" target ="_blank">https://github.com/simonom/PyCHAM</a>
**Version:** 0.0.5
**Reviewers:** dhhagan, andreas-h
