### 2019-11-08 #1879 @kyleniemeyer
Hi @simonom, thanks for your interest in JOSS. Normally at this stage I would assign an editor, who would then find reviewers in this pre-review issue.

However, I noticed some major issues with your submission that need to be addressed (which any reviewers would likely get stuck on):
- Your software needs an open-source license, from this [list](https://opensource.org/licenses)
- You've created a Python package, but it isn't set up like a package that can be installed in the standard ways. This isn't too much work, but it goes a long way towards making your software usable by others. There are multiple guides out there on doing this: https://docs.python-guide.org/shipping/packaging/
