### 2019-11-14 cf71406
modified to be suitable for pip install
