+++ PyCHAM/Unit_Testing/Rate_coefficients.py
@@ -1,174 +1,0 @@
-'''module for calculating gas-phase reaction rate coefficients, automatically generated by eqn_parser'''
-
-##################################################################################################### 
-# Python function to hold expressions for calculating rate coefficients for a given equation number # 
-#    Copyright (C) 2017  David Topping : david.topping@manchester.ac.uk                             # 
-#                                      : davetopp80@gmail.com                                       # 
-#    Personal website: davetoppingsci.com                                                           # 
-#                                                                                                   # 
-#                                                                                                   # 
-#                                                                                                   # 
-##################################################################################################### 
-# Minor modified by XSX
-# File Created as 2019-11-01 15:49:07.001737
-
-import numpy
-
-def evaluate_rates(RO2, H2O, TEMP, lightm):
-    # mcm_constants_dict: given by mcm_constants.py
-    # RO2: specified by the chemical scheme. eg: subset of MCM
-    # H2O, TEMP: given by the user
-    # lightm: given by the user and is 0 for lights off and 1 for on
-    # Creating reference to constant values used in rate expressions
-    # mcm_constants_dict: given by MCM_constants.py
-    KRO2NO = 9.031312353614072e-12 
-    KRO2HO2 = 2.277786778370343e-11 
-    KAPHO2 = 1.3915587832143401e-11 
-    KAPNO = 1.9837375038963997e-11 
-    KRO2NO3 = 2.3e-12 
-    KNO3AL = 2.733975964953383e-15 
-    KDEC = 1000000.0 
-    KROPRIM = 9.140096052746274e-15 
-    KROSEC = 9.140096052746274e-15 
-    KCH3O2 = 3.5035432056788604e-13 
-    K298CH3O2 = 3.5e-13 
-    KD0 = 0.2858654787161464 
-    KDI = 0.0003865354977050049 
-    KRD = 739.5581529081514 
-    FCD = 0.3 
-    NCD = 1.4140560065060288 
-    FD = 0.7903214205531268 
-    KBPAN = 0.0003050747741265704 
-    KC0 = 6.944423033246356e-09 
-    KCI = 1.2066992504885919e-11 
-    KRC = 575.4891312341963 
-    FCC = 0.3 
-    NC = 1.4140560065060288 
-    FC = 0.7785522770353273 
-    KFPAN = 9.378487940710399e-12 
-    K10 = 2.4859762715193917e-12 
-    K1I = 2.994437979215886e-11 
-    KR1 = 0.08301979499239324 
-    FC1 = 0.85 
-    NC1 = 0.8396379643428482 
-    F1 = 0.9406666674267817 
-    KMT01 = 2.1592172418682586e-12 
-    K20 = 3.2297706766030638e-12 
-    K2I = 2.2965879943120184e-11 
-    KR2 = 0.1406334390235544 
-    FC2 = 0.6 
-    NC2 = 1.0317479120127726 
-    F2 = 0.7380514176422049 
-    KMT02 = 2.0898360024993884e-12 
-    K30 = 9.088989095412873e-11 
-    K3I = 1.8976508649648676e-12 
[…]
