+++ PyCHAM/Unit_Testing/test_kimt_prep.py
@@ -36,7 +36,7 @@ if surfT!=72.0:
 	print('issue with surfT')
 if int(Cw*1e-13)!= 301:
 	print('issue with Cw')
-if int(kgwt*1e25)!= 332:
+if int(kgwt*1e13)!= 332:
 	print('issue with kgwt')
 
 print('if no issues printed above, code is fine')

[…]
