### 2019-11-14 fb9fe99
modified to be suitable for pip install
