### 2019-11-14 #1879 @simonom
Many thanks @kyleniemeyer ,

These issues you raised (inclusion of license and availability on PyPi) have now been addressed.

Regards,
Simon O'Meara
