+++ build/lib/PyCHAM/PhotolysisRates.py
@@ -1,0 +1,82 @@
+'''module for calculating photolysis rates for MCM (Master Chemical Mechanism) equations'''
+
+# Originaated from KPP (Kinetic Pre-processor)
+
+import scipy
+import numpy
+import ipdb
+
+def PhotolysisCalculation(time, lat, lon):
+
+	(theta, secx, cosx) = zenith(time, lat, lon)
+
+	J = [0.0 for i in range(62)] # modified, so that we dont have to check None type
+    
+    #J          L           M          N
+	J[1]=6.073E-05*cosx**(1.743)*numpy.exp(-1.0*0.474*secx)
+	J[2]=4.775E-04*cosx**(0.298)*numpy.exp(-1.0*0.080*secx)
+	J[3]=1.041E-05*cosx**(0.723)*numpy.exp(-1.0*0.279*secx)
+	J[4]=1.165E-02*cosx**(0.244)*numpy.exp(-1.0*0.267*secx)
+	J[5]=2.485E-02*cosx**(0.168)*numpy.exp(-1.0*0.108*secx)
+	J[6]=1.747E-01*cosx**(0.155)*numpy.exp(-1.0*0.125*secx)
+	J[7]=2.644E-03*cosx**(0.261)*numpy.exp(-1.0*0.288*secx)
+	J[8]=9.312E-07*cosx**(1.230)*numpy.exp(-1.0*0.307*secx)
+	J[11]=4.642E-05*cosx**(0.762)*numpy.exp(-1.0*0.353*secx)
+	J[12]=6.853E-05*cosx**(0.477)*numpy.exp(-1.0*0.323*secx)
+	J[13]=7.344E-06*cosx**(1.202)*numpy.exp(-1.0*0.417*secx)
+	J[14]=2.879E-05*cosx**(1.067)*numpy.exp(-1.0*0.358*secx)
+	J[15]=2.792E-05*cosx**(0.805)*numpy.exp(-1.0*0.338*secx)
+	J[16]=1.675E-05*cosx**(0.805)*numpy.exp(-1.0*0.338*secx)
+	J[17]=7.914E-05*cosx**(0.764)*numpy.exp(-1.0*0.364*secx)
+	J[18]=1.140E-05*cosx**(0.396)*numpy.exp(-1.0*0.298*secx)
+	J[19]=1.140E-05*cosx**(0.396)*numpy.exp(-1.0*0.298*secx)
+	J[21]=7.992E-07*cosx**(1.578)*numpy.exp(-1.0*0.271*secx)
+	J[22]=5.804E-06*cosx**(1.092)*numpy.exp(-1.0*0.377*secx)
+	J[23]=1.836E-05*cosx**(0.395)*numpy.exp(-1.0*0.296*secx)
+	J[24]=1.836E-05*cosx**(0.395)*numpy.exp(-1.0*0.296*secx)
+	J[31]=6.845E-05*cosx**(0.130)*numpy.exp(-1.0*0.201*secx)
+	J[32]=1.032E-05*cosx**(0.130)*numpy.exp(-1.0*0.201*secx)
+	J[33]=3.802E-05*cosx**(0.644)*numpy.exp(-1.0*0.312*secx)
+	J[34]=1.537E-04*cosx**(0.170)*numpy.exp(-1.0*0.208*secx)
+	J[35]=3.326E-04*cosx**(0.148)*numpy.exp(-1.0*0.215*secx)
+	J[41]=7.649E-06*cosx**(0.682)*numpy.exp(-1.0*0.279*secx)
+	J[51]=1.588E-06*cosx**(1.154)*numpy.exp(-1.0*0.318*secx)
[…]
