### 2019-11-14 #1879 @labarba
@simonom — please check in your paper the unresolved references (shown in the PDF as ???).
