### 2019-11-14 #1879 @labarba
@jedbrown — Could you handle this submission as editor? (Lindsey is stepping down from the board to focus on her job search and we're low on geosciences editors now!)
