### 2019-11-15 #1879 @jedbrown
@simonom Welcome! Could you please check your paper for **???** references, and also consider more explicitly laying out the problems that are solved (e.g., can you write the structural form of the differential equations?) and related work (even if not written in Python)? Also, I would really encourage you to add continuous integration (e.g., Travis-CI or Azure DevOps) to run and report on your test suite, and to use formatting in your README (lots of commands and path literals don't use `verbatim`).
