+++ paper.md
@@ -40,9 +40,9 @@ bibliography: paper_refs.bib
 
 # Summary
 
-PyCHAM (CHemistry with Aerosol Microphysics in Python) is an open-access O-D box model for environmental chamber studies.  Environmental chambers provide a means for atmospheric scientists to interrogate aerosol processes [@Finlayson:2000, @Schwantes:2017, @Hidy:2019].  Allying PyCHAM with chamber measurements allows quantification of unknown parameters, such as branching ratios for oxidation schemes [@Chen:2005].  Although several box models have already been published [@Pierce:2008, @Roldin:2014, @Sunol:2018, @Roldin:2019], PyCHAM is novel in its ease of accessibility and utility.  The intention therefore, is that it will be readily employed by research groups undertaking environmental chamber measurements.
+PyCHAM (CHemistry with Aerosol Microphysics in Python) is an open-access O-D box model for environmental chamber studies.  Environmental chambers provide a means for atmospheric scientists to interrogate aerosol processes [@Finlayson:2000; @Schwantes:2017; @Hidy:2019].  Allying PyCHAM with chamber measurements allows quantification of unknown parameters, such as branching ratios for oxidation schemes [@Chen:2005].  Although several box models have already been published [@Pierce:2008; @Roldin:2014; @Sunol:2018; @Roldin:2019], PyCHAM is novel in its ease of accessibility and utility.  The intention therefore, is that it will be readily employed by research groups undertaking environmental chamber measurements.
 
-With air quality and climate models increasingly important to guiding sustainable societies, the accuracy of simulations must suffice [e.g. @Tong:2019].  However, research shows that the simulated aerosol effects in these models provides a relatively high amount of uncertainty [@Johnson:2018].  The combination of box models like PyCHAM with environmental chamber measurements to better constrain aerosol processes is therefore necessary to ultimately improve societal sustainability.
+With air quality and climate models increasingly important to guiding sustainable societies, the accuracy of simulations must suffice [@Tong:2019].  However, research shows that the simulated aerosol effects in these models provides a relatively high amount of uncertainty [@Johnson:2018].  The combination of box models like PyCHAM with environmental chamber measurements to better constrain aerosol processes is therefore necessary to ultimately improve societal sustainability.
 
 Funding for model development has been provided by the EUROCHAMP-2020 research project [@EUROCHAMP:2020].  At the time of writing, PyCHAM is being used to investigate the autoxidation of organic vapours in the atmosphere and the partitioning of semi-volatile vapours to soot aerosol.  Here is the repository: github.com/simonom/PyCHAM.
 
