+++ paper.bib
@@ -1,97 +1,97 @@
-@book{Finlayson2000,
-    author    = "B.J. Finlayson-Pitts and J.N.Jr. Pitts",
-    title     = "Chemistry of the upper and lower atmosphere: Theory, experiments and applications",
-    publisher = "Academic Press",    
-    year = "2000",
-    edition = "1",
-    address = "San Diego" 
+@book{Finlayson:2000,
+    author    = {B.J. Finlayson-Pitts and J.N.Jr. Pitts},
+    title     = {Chemistry of the upper and lower atmosphere: Theory, experiments and applications},
+    publisher = {Academic Press},    
+    year = {2000},
+    edition = {1},
+    address = {San Diego} 
 }
 
-@book{Schwantes2017,
-    author    = "R.H. Schwantes and R.C. McVay and X. Zhang and M.M. Coggon and H. Lignell and R.C. Flagan and P.O. Wennberg and J.H. Seinfeld",
-    title     = "Advances in Atmospheric Chemistry, Chapter 1: Science of the Environmental Chamber",
-    publisher = "World Scientific",    
-    year = "2017",
-    doi = "10.1142/10216" 
+@book{Schwantes:2017,
+    author    = {R.H. Schwantes and R.C. McVay and X. Zhang and M.M. Coggon and H. Lignell and R.C. Flagan and P.O. Wennberg and J.H. Seinfeld},
+    title     = {Advances in Atmospheric Chemistry, Chapter 1: Science of the Environmental Chamber},
+    publisher = {World Scientific},    
+    year = {2017},
+    doi = {10.1142/10216} 
 }
 
-@article{Hidy2019,    
-	author = "G.M. Hidy",
-	title = "Atmospheric Chemistry in a Box or a Bag",
-	journal = "Atmosphere-Basel",
-	year = "2019",
-	volume = "10(7)",
-	doi = "10.3390/atmos10070401",
+@article{Hidy:2019,    
+	author = {G.M. Hidy},
+	title = {Atmospheric Chemistry in a Box or a Bag},
+	journal = {Atmosphere-Basel},
+	year = {2019},
+	volume = {10(7)},
+	doi = {10.3390/atmos10070401},
     }
 
-@article{Chen2005,    
-	author = "J. Chen and R.J. Griffin",
-	title = "Modeling secondary organic aerosol formation from oxidation of $\alpha$-pinene, $\beta$-pinene and $d$-limonene",
-	journal = "Atmos. Env.",
-	year = "2005",
-	volume = "39",
-	pages = "7731-7744",
-	doi = "10.1016/j.atmosenv.2005.05.049",
+@article{Chen:2005,    
+	author = {J. Chen and R.J. Griffin},
+	title = {Modeling secondary organic aerosol formation from oxidation of $\alpha$-pinene, $\beta$-pinene and $d$-limonene},
+	journal = {Atmos. Env.},
+	year = {2005},
+	volume = {39},
+	pages = {7731-7744},
+	doi = {10.1016/j.atmosenv.2005.05.049},
     }
 
-@article{Pierce2008,    
-	author = "J. Pierce and G. Englehart and L. Hildebrandt and E. Weitkamp and E. Parthak and N. Donahue",
-	title = "Constraining Particle Evolution from Wall Losses, Coagulation and Condensation-Evaporation in Smog Chamber Experiments: Optimal Based Size Distribution Measurements",
-	journal = "Aerosol Sci. Tech.",
-	year = "2008",
-	volume = "42",
-	doi = "10.1080/02786820802389251",
+@article{Pierce:2008,    
[…]
