### 2019-11-15 0c30c98
improved directions in README
