### 2019-11-19 0db6a75
updated to work with 0 size bins
