+++ build/lib/PyCHAM/__init__.py
@@ -1,0 +1,1 @@
+'''Used to tell setup.py that PyCHAM is a package (when creating dist and build folders), but is empty'''
