+++ build/lib/PyCHAM/eqn_parser.py
@@ -1,0 +1,480 @@
+''' module for parsing the equation file, including converting MCM names of components into SMILES strings - see the README for how to format the equation and xml files'''
+
+import os
+import re
+import collections
+import sys
+import datetime
+import numpy as np
+import pybel
+import formatting
+import xmltodict # for opening and converting xml files to python dictionaries
+import ipdb
+
+# ----------Extraction of eqn info----------
+# Extract the mechanism information
+def extract_mechanism(filename, xmlname, TEMP, PInit, Comp0, testf):
+
+	# inputs:
+	# testf - flag for operating in normal mode (0) or testing mode (1)
+	
+	if testf == 1:
+		return(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)
+    	
+	# Open the file
+	f_open_eqn = open(filename, mode='r')
+	
+	# read the file and store everything into a list
+	# reaction rates have units /s
+	total_list_eqn = f_open_eqn.readlines()
+	f_open_eqn.close()
+	if (f_open_eqn.closed == False):
+		print('IOError')
+		print('Eqn file not closed')
+		sys.exit()
+
+	# equation list without comments
+	naked_list_eqn = formatting.remove_comments(total_list_eqn)
+	
+	# calculate gas-phase concentrations of M, N2 and O2 (molecules/cc (air))
+	# 1.0e-6 converts from molecules/m3 to molecules/cc
+	M_val = (PInit/(8.3144598*TEMP)*6.0221409e+23)*1.0e-6
+	N2_val = M_val*0.79
+	O2_val = M_val*0.2096
+			
+	# format the equation list
+	
+	# naked_list_eqn contains everything except the comments starting with //
+	print('Now parsing the eqn info...\n')
+	num_eqn = len(naked_list_eqn)
+	
+	
+	# --open and initialise the xml file for converting chemical names to SMILES-----
+	with open(xmlname) as fd:
+		doc = xmltodict.parse(fd.read())
+
+	a = doc['mechanism']['species_defs']['species']
+	spec_numb = list(('0',) * len(a))
+	spec_name = list(('0',) * len(a))
+	spec_smil = list(('0',) * len(a))
+	
+	for i in range(len(a)):
+		spec_numb[i] = a[i]['@species_number']
+		spec_name[i] = a[i]['@species_name']
+		if "smiles" in a[i]:
+			spec_smil[i] = a[i]['smiles']
+		elif spec_name[i][0]=='O' or spec_name[i][0]=='H':
+			 spec_smil[i] = '['+spec_name[i]+']'
+		else:
+			 spec_smil[i] = spec_name[i] 
+    
+	species_step = 0 # log the number of unique species
+	max_no_reac = 0.0 # log maximum number of reactants in a reaction
+	max_no_prod = 0.0 # log maximum number of products in a reaction
+	
+	# convert input component names for components present in gas phase at experiment
+	# start from chemical scheme names to SMILES
+	init_SMIL = []
+	
+	for species_step in range(len(Comp0)): 
+		name_indx = spec_name.index(Comp0[species_step])
+		init_SMIL.append(spec_smil[name_indx])
+	species_step = 0 # ready for equation loop
+	
+	# initialising lists
+	
+	# matrix to record indices of reactants (cols) in each equation (rows)
+	rindx = np.zeros((num_eqn, 1)).astype(int)
+	# matrix to record indices of products (cols) in each equation (rows)
[…]

[…]
