+++ build/lib/PyCHAM/Rate_coeffs.py
@@ -10,7 +10,7 @@
 #                                                                                                   # 
 ##################################################################################################### 
 # Minor modified by XSX
-# File Created as 2019-11-14 10:07:55.065063
+# File Created as 2019-11-19 14:47:02.031647
 
 import numpy
 
@@ -21,152 +21,152 @@ def evaluate_rates(RO2, H2O, TEMP, lightm):
     # lightm: given by the user and is 0 for lights off and 1 for on
     # Creating reference to constant values used in rate expressions
     # mcm_constants_dict: given by MCM_constants.py
-    KRO2NO = 9.180825715586685e-12 
-    KRO2HO2 = 2.416926637791736e-11 
-    KAPHO2 = 1.4551688760645419e-11 
-    KAPNO = 2.0101502680038634e-11 
+    KRO2NO = 9.031312353614072e-12 
+    KRO2HO2 = 2.277786778370343e-11 
+    KAPHO2 = 1.3915587832143401e-11 
+    KAPNO = 1.9837375038963997e-11 
     KRO2NO3 = 2.3e-12 
-    KNO3AL = 2.5116077328102446e-15 
+    KNO3AL = 2.733975964953383e-15 
     KDEC = 1000000.0 
-    KROPRIM = 9.015884918617853e-15 
-    KROSEC = 9.015884918617853e-15 
-    KCH3O2 = 3.5623566538021364e-13 
+    KROPRIM = 9.140096052746274e-15 
+    KROSEC = 9.140096052746274e-15 
+    KCH3O2 = 3.5035432056788604e-13 
     K298CH3O2 = 3.5e-13 
-    KD0 = 0.1613839467444603 
-    KDI = 0.00020570494794194444 
-    KRD = 784.5409085152743 
+    KD0 = 0.2764847462539585 
+    KDI = 0.0003865354977050049 
+    KRD = 715.2894052306818 
     FCD = 0.3 
     NCD = 1.4140560065060288 
-    FD = 0.7929743154543676 
-    KBPAN = 0.00016291108878962242 
-    KC0 = 7.493072176355027e-09 
-    KCI = 1.2214576331165244e-11 
-    KRC = 613.4533014654471 
+    FD = 0.7888029188404472 
+    KBPAN = 0.0003044746624302591 
+    KC0 = 6.716540412120828e-09 
+    KCI = 1.2066992504885919e-11 
+    KRC = 556.6043410900689 
     FCC = 0.3 
     NC = 1.4140560065060288 
-    FC = 0.7816282619662761 
-    KFPAN = 9.531720259332986e-12 
-    K10 = 2.4903359658729046e-12 
-    K1I = 2.982328880018226e-11 
-    KR1 = 0.08350306307792872 
+    FC = 0.776923817877948 
+    KFPAN = 9.358320650293402e-12 
+    K10 = 2.4043984664090836e-12 
+    K1I = 2.994437979215886e-11 
+    KR1 = 0.0802954839304667 
     FC1 = 0.85 
     NC1 = 0.8396379643428482 
-    F1 = 0.9404990169661523 
-    KMT01 = 2.1616538131101394e-12 
-    K20 = 3.2310676634020842e-12 
-    K2I = 2.2891553138312854e-11 
-    KR2 = 0.14114672097081743 
+    F1 = 0.9416197104271332 
+    KMT01 = 2.0957497475174937e-12 
[…]
