+++ paper.md
@@ -48,8 +48,10 @@ Funding for model development has been provided by the EUROCHAMP-2020 research p
 
 The model employs non-equilibrium equations to simulate the known processes occurring in environmental chambers.  At its core is integration of ordinary differential equations for gas-phase photochemistry and gas partitioning to particles and walls.  The general equation for chemical reactions is: 
 
-$$\frac{d[i]}{dt} = \pm k_{n}[a]^{a_{s}}[b]^{b_{s}}
+$$\frac{d[i]}{dt} = \pm k_{n}[a]^{a_{s}}[b]^{b_{s}},
 $$
+
+where $$i$$ is the affected species.
 
 
 particle loss to walls, coagulation and nucleation.  It takes a sectional approach to particulates, dividing particles into a number of size bins and treating their changing size using the moving-centre approach [@Jacobson:2005].  It builds upon PyBOX [@Topping:2018] which did not include coagulation, nucleation or a sectional method.
