### 2019-11-19 cd99833
Update paper.md
