+++ paper.bib
@@ -127,3 +127,13 @@
 	volume = {116},
 	doi = {http://dx.doi.org/10.1016/j.matcom.2015.04.007},
     }
+    
+@article{Zhang:2015,
+	author = {X. Zhang and R.H. Schwantes and R.C. McVay and H. Lignell and M.M. Coggon and R.C. Flagan and J.H. Seinfeld},
+	title = {Vapor wall deposition in Teflon chambers},
+	year = {2015},
+	journal = {Atmos. Chem. Phys.},
+	volume = {15},
+	pages={4197-4214},
+	doi={10.5194/acp-15-4197-2015},
+}
