### 2019-11-19 daa1d63
More explicit equations and application
