+++ paper.bib
@@ -137,3 +137,13 @@
 	pages={4197-4214},
 	doi={10.5194/acp-15-4197-2015},
 }
+
+@article{Ehn:2014,
+	author = {M. Ehn and J.A. Thornton and E. Kleist and M. Sipil{\"a} and H. Junninen and I. Pullinen and M. Springer and F. Rubach and R. Tillmann and B. Lee and F. Lopez-Hilfiker and S. Andres and I. Acir and M. Rissanen and T. Jokinen and S. Schobesberger and J. Kangasluoma and J. Kontkanen and T. Nieminen and T. Kurt{\'e}n and L. B. Nielsen and S. J{\o}rgensen and H.G. Kjaergaard and M. Canagaratna and M.D. Maso and T. Berndt and T. Pet{\"a}j{\"a} and A. Wahner and V. Kerminen and M. Kulmala and D.R. Worsnop and J. Wildt and T.F. Mentel},
+	title = {A large source of low-volatility secondary organic aerosol},
+	year = {2014},
+	journal = {Nature},
+	volume = {506},
+	pages = {476--479},
+	doi={10.1038/nature13032},
+}
