+++ README.md
@@ -1,4 +1,4 @@
-**PyCHAM: Python CHemistry with Aerosol Microphysics Box Model**
+# PyCHAM: Python CHemistry with Aerosol Microphysics Box Model
 
 Welcome to the PyCHAM.  Funding has been provided from the [EUROCHAMP-2020 research project](http://www.eurochamp.org).  Please contact Simon O'Meara (simon.omeara@manchester.ac.uk) with any issues, comments or suggestions.
 
@@ -7,7 +7,7 @@ PyCHAM is an open-access computer code (written in Python) for simulating aeroso
 The programme is controlled via a graphical user interface (GUI).  There are two options for installing, via conda and via pip. 
 
 
-**conda Install**
+# conda Install
 
 1. Download the repository from github.com/simonom/PyCHAM
 
@@ -19,7 +19,7 @@ The programme is controlled via a graphical user interface (GUI).  There are two
 
 5. Install complete
 
-**pip Install**
+# pip Install
 
 1. Inside the terminal/command prompt, create a virtual env using venv: python -m venv PyCHAM
 
@@ -33,7 +33,7 @@ The programme is controlled via a graphical user interface (GUI).  There are two
 
 6. Install complete
 
-**Once Installed**
+# Once Installed
 
 1. For model inputs, ensure you have: a .txt file chemical reaction scheme, a .xml file for converting species names to SMILE strings and a .txt file stating values of model variables (e.g. temperature) - see details about these three files below
 
@@ -45,16 +45,16 @@ The programme is controlled via a graphical user interface (GUI).  There are two
 
 5. The 'plot results' button produces (and saves in the output folder) two figures, the first called contours.png that shows the particle number distribution, SOA mass and particle number concentration against time, and the second called gas_pbb.png that shows the gas-phase concentrations of initial components with time  
 
-**Unit Testing and Example Run**
+# Unit Testing and Example Run
 
 Unit tests for PyCHAM modules can be found in the PyCHAM/Unit_Testing folder in the Github repository.  To use, cd to this folder and use python test_module.py with module replaced by the name of the module to be tested.
 
 The example run output is saved in output/Example_Run/Example_output in the Github repository.  To reproduce this (e.g. for testing), use the inputs/Example_Run.txt for the chemical scheme, inputs/Example_Run_xml.xml for the xml file and inputs/Example_Run_inputs.txt for the model inputs.  Note that the example output was produced using version 0.2.1 of PyCHAM.
 
-**Chemical Scheme .txt file**
+# Chemical Scheme .txt file
 
 An example chemical scheme .txt file is given in the inputs folder (of the Github repository), called 'Example_Run.txt', which has been obtained
-from the Master Chemical Mechanism website and modified.
+from the [Master Chemical Mechanism website](http://mcm.leeds.ac.uk/MCM/) and modified.
 
 // can be used for preceding comments, as in the example.
 
@@ -66,14 +66,14 @@ The rate coefficients must adhere to the following rules:
[…]
