+++ build/lib/PyCHAM/front.py
@@ -1,136 +1,0 @@
-'''module called on by __main__ to initiate the box model'''
-
-import numpy as np
-from ode_gen import ode_gen
-import matplotlib.pyplot as plt
-import os
-import eqn_parser # to parse the .scm file
-from init_conc_func import init_conc_func
-from pp_intro import pp_intro
-from kimt_prep import kimt_prep
-from saving import saving
-import time # timing how long operations take
-import user_input as ui
-import pickle # for storing inputs
-from volat_calc import volat_calc
-
-def run(testf):
-	
-	# inputs:
-	# testf - test flag, 1 for test mode (called by test_front.py), 2 for test mode 
-	# (called by test_PyCHAM.py), 0 for normal mode
-	
-	if testf==2:
-		print('"Run Model" button works fine')
-		return()
-	if testf==1:
-		print('calling user_input.py')
-	# module to ask, receive and return required inputs
-	[fname, num_sb, lowersize, uppersize, end_sim_time, resfname, tstep_len, 
-	tmax, TEMP, PInit, RH, lat, lon, start_sim_time, save_step, Cw, 
-	ChamR, nucv1, nucv2, nucv3, nuc_comp, inflectDp, pwl_xpre,  
-	pwl_xpro, inflectk, xmlname, init_conc, Comp0, Rader, voli, volP, 
-	pconc, std, loc, scale, core_diss, light_stat, light_time, kgwt, testm] = ui.run(0, testf)
-	
-	if testm == 1:
-		print('PyCHAM calls front fine, now returning to PyCHAM.py')
-		print('Please select the "Plot Results" button to ensure this works fine')
-		return()
-	
-	if testf==1:
-		print('user_input.py called and returned fine')
-		print('calling eqn_parser.extract_mechanism')
-	
-	# obtain gas-phase reaction info
-	[rindx, pindx, rstoi, pstoi, reac_coef, spec_list, Pybel_objects, num_eqn, num_speci, 
-		RO2_indices, nreac, nprod, prodn, 
-		reacn, M_val, N2_val, O2_val, 
-		init_SMIL] = eqn_parser.extract_mechanism(fname, xmlname, 
-							TEMP, PInit, Comp0, testf)
-	
-	if testf==1:
-		print('eqn_parser.extract_mechanism called and returned fine')
-		print('calling init_conc_func')
-	# set up initial gas-phase concentration array
-	[y, H2Oi, Psat_water, y_mw, num_speci, 
-	Cfactor, y_indx_plot, corei] = init_conc_func(num_speci, init_SMIL, spec_list, 
-							init_conc, TEMP, RH, M_val, N2_val, O2_val, reac_coef, fname, 
-							PInit, start_sim_time, lat, lon, Pybel_objects, testf, pconc)
-	
-	
-	if testf==1:
-		print('init_conc_func called and returned fine')
-		print('calling kimt_prep')
-	# set up partitioning variables
-	[DStar_org, mfp, accom_coeff, therm_sp, surfT, Cw, kgwt] = kimt_prep(y_mw, TEMP, 
-														num_speci, testf, Cw, kgwt)
-	
-	# volatility (molecules/cc (air)) and density (rho, kg/m3) of components
-	if testf==1:
-		print('kimt_prep called and returned fine')
-		print('calling volat_calc.py')
-	[Psat, y_dens, Psat_Pa] = volat_calc(spec_list, Pybel_objects, TEMP, H2Oi, num_speci,  
-								Psat_water, voli, volP, testf, corei)
-	
-	if testf==1:
-		print('volat_calc called and returned fine')
-		print('calling wall_prep')
[…]
