### 2019-11-20 #1879 @simonom
And now the improved description of ode equations and related work has been included in the paper.  In addition, the README has been formatted @jedbrown.
