### 2019-11-20 e9049a6
added install commands
