### 2019-11-20 96599c5
compatible with conda
https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/use-conda-with-travis-ci.html
