+++ .travis.yml
@@ -1,99 +1,25 @@
 language: python
+python:
+  -"3.6"
+install:
+  - sudo apt-get update
+  # We do this conditionally because it saves us some downloading if the
+  # version is the same.
+  - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
+      wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
+    else
+      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+    fi
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - source "$HOME/miniconda/etc/profile.d/conda.sh"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
+  # Useful for debugging any issues with conda
+  - conda info -a
 
-channels:
-  - openbabel
-  - numba
-  - conda-forge
-  - anaconda
-  - defaults
-before_install:
-  - sudo apt-get -y install assimulo=3.0=py37hc04db61_1007
-  - sudo apt-get -y install backcall=0.1.0=py_0
-  - sudo apt-get -y install bzip2=1.0.8=h516909a_0
-  - sudo apt-get -y install ca-certificates=2019.10.16=0
-  - sudo apt-get -y install cairo=1.14.12=h8948797_3
-  - sudo apt-get -y install certifi=2019.9.11=py37_0
-  - sudo apt-get -y install cycler=0.10.0=py_1
-  - sudo apt-get -y install dbus=1.13.6=h746ee38_0
-  - sudo apt-get -y install decorator=4.4.0=py_0
-  - sudo apt-get -y install expat=2.2.6=he6710b0_0
-  - sudo apt-get -y install fontconfig=2.13.0=h9420a91_0
-  - sudo apt-get -y install freetype=2.10.0=he983fc9_1
-  - sudo apt-get -y install gettext=0.19.8.1=hc5be6a0_1002
-  - sudo apt-get -y install git=2.20.1=pl526hacde149_0
-  - sudo apt-get -y install gitdb2=2.0.6=py_0
-  - sudo apt-get -y install gitpython=3.0.4=py_0
-  - sudo apt-get -y install glib=2.56.2=hd408876_0
-  - sudo apt-get -y install gst-plugins-base=1.14.0=hbbd80ab_1
-  - sudo apt-get -y install gstreamer=1.14.0=hb453b48_1
-  - sudo apt-get -y install icu=58.2=h9c2bf20_1
-  - sudo apt-get -y install ipdb=0.12.2=py_0
-  - sudo apt-get -y install ipython=7.7.0=py37h5ca1d4c_0
-  - sudo apt-get -y install ipython_genutils=0.2.0=py_1
-  - sudo apt-get -y install jedi=0.15.1=py37_0
-  - sudo apt-get -y install jpeg=9c=h14c3975_1001
-  - sudo apt-get -y install kiwisolver=1.1.0=py37hc9558a2_0
-  - sudo apt-get -y install krb5=1.16.1=h173b8e3_7
-  - sudo apt-get -y install libblas=3.8.0=12_openblas
-  - sudo apt-get -y install libcblas=3.8.0=12_openblas
[…]
