+++ .travis.yml
@@ -1,6 +1,6 @@
 language: python
 python:
-  -"3.6"
+  -"3.6.7"
 install:
   - sudo apt-get update
   # We do this conditionally because it saves us some downloading if the
