+++ .travis.yml
@@ -13,12 +13,13 @@ install:
   - source "$HOME/miniconda/etc/profile.d/conda.sh"
   - hash -r
   - conda config --set always_yes yes --set changeps1 no
+  - conda config --add channels conda-forge
   - conda update -q conda
   # Useful for debugging any issues with conda
   - conda info -a
 
   # Replace dep1 dep2 ... with your dependencies
-  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION _libgcc_mutex=0.1=main assimulo backcall=0.1.0=py_0 bzip2 ca-certificates=2019.10.16=0 cairo=1.14.12=h8948797_3 certifi=2019.9.11=py37_0 cycler=0.10.0=py_1 dbus=1.13.6=h746ee38_0 decorator=4.4.0=py_0 expat=2.2.6=he6710b0_0 fontconfig=2.13.0=h9420a91_0 freetype=2.10.0=he983fc9_1 gettext git=2.20.1=pl526hacde149_0 gitdb2=2.0.6=py_0 gitpython=3.0.4=py_0 glib=2.56.2=hd408876_0 gst-plugins-base=1.14.0=hbbd80ab_1 gstreamer=1.14.0=hb453b48_1 icu=58.2=h9c2bf20_1 ipdb=0.12.2=py_0 ipython=7.7.0=py37h5ca1d4c_0 ipython_genutils=0.2.0=py_1 jedi=0.15.1=py37_0 jpeg=9c=h14c3975_1001 kiwisolver krb5=1.16.1=h173b8e3_7 libblas=3.8.0=12_openblas libcblas=3.8.0=12_openblas libcurl=7.65.3=h20c2e04_0 libedit=3.1.20181209=hc058e9b_0 libffi=3.2.1=he1b5a44_1006 libgcc=7.2.0=h69d50b8_2 libgcc-ng=9.1.0=hdf63c60_0 libgfortran-ng=7.3.0=hdf63c60_0 libiconv=1.15=h516909a_1005 liblapack=3.8.0=12_openblas libopenblas=0.3.7=h6e990d7_1 libpng libssh2=1.8.2=h1ba5d50_0 libstdcxx-ng=9.1.0=hdf63c60_0 libuuid=1.0.3=h1bed415_2 libxcb=1.13=h14c3975_1002 libxml2=2.9.9=hea5a465_1 llvmlite=0.29.0=py37hf484d3e_0 matplotlib=3.1.0=py37h5429711_0 metis=5.1.0=hf484d3e_1003 ncurses=6.1=hf484d3e_1002 numba=0.45.1=py37h962f231_0 numpy=1.17.0=py37h95a1406_0 openbabel=2.4.1=py37_5 openssl pandas=0.25.0=py37he6710b0_0 parso=0.5.1=py_0 pcre=8.43=he6710b0_0 perl=5.26.2=h14c3975_0 pexpect=4.7.0=py37_0 pickleshare=0.7.5=py37_1000 pip=19.2.2=py37_0 pixman=0.38.0=h7b6447c_0 prompt_toolkit pthread-stubs ptyprocess pygments=2.4.2=py_0 pyparsing=2.4.2=py_0 pyqt=5.9.2=py37h05f1152_2 python-dateutil pytz=2019.2=py_0 qt=5.9.7=h5867ecd_1 readline=8.0=hf8c457e_0 scipy setuptools=41.2.0=py37_0 sip six smmap2=2.0.5=py_0 sqlite suitesparse sundials=3.2.1=h775d46b_0 tbb=2019.7=hc9558a2_0 tk=8.6.9=hed695b0_1002 tornado traitlets wcwidth=0.1.
[…]
