### 2019-11-20 6dff069
added PyCHAM file for use in Travis CI
