### 2019-11-20 15ff6ff
changed script in travis yml file
