+++ .travis.yml
@@ -24,4 +24,4 @@ install:
   
   - conda activate test-environment
   - python setup.py install
-script: python PyCHAM
+script: python PyCHAM/PyCHAM_CI_test.py
