### 2019-11-20 #1879 @simonom
Finally, continuous integration using Travis-CI has been implemented, thanks @jedbrown .
