### 2019-11-20 070bb9b
allows empty inputs
