+++ PyCHAM/Rate_coeffs.py
@@ -10,7 +10,7 @@
 #                                                                                                   # 
 ##################################################################################################### 
 # Minor modified by XSX
-# File Created as 2019-11-20 11:44:32.510435
+# File Created as 2019-11-20 16:45:34.173610
 
 import numpy
 
