+++ PyCHAM/PyCHAM.py
@@ -98,61 +98,148 @@ class PyCHAM(QWidget):
 			key, value = in_list[i].split('=')
 			key = key.strip() # a string
 			if key == 'Res_file_name':
+				if value.split(',')==['\n']:
+					print('Error: no requested results file name detected in model inputs file, please supply')
+					sys.exit()
 				resfname = str(value.strip())
 			if key == 'Total_model_time':
-				end_sim_time = float(value.strip())
+				if value.split(',')==['\n']:
+					print('Error: no total model run time detected in model inputs file, please supply')
+					sys.exit()
+				else:
+					end_sim_time = float(value.strip())
 			if key == 'Time_step':
-				tstep_len = float(value.strip())
+				if value.split(',')==['\n']:
+					print('Notice: No model time step detected in model inputs file, defaulting to 60s')
+					tstep_len = float(60.0)
+				else:
+					tstep_len = float(value.strip())
 			if key == 'Recording_time_step':
-				save_step = float(value.strip())
+				if value.split(',')==['\n']:
+					print('Notice: no recording time step detected in model inputs file, defaulting to 60s')
+					save_step = float(60.0)
+				else:
+					save_step = float(value.strip())
 			if key == 'Number_size_bins':
-				num_sb = int(value.strip())
+				if value.split(',')==['\n']:
+					print('Notice: no size bin number detected in model inputs, defaulting to zero')
+					num_sb = int(0)
+				else:
+					num_sb = int(value.strip())
 			if key == 'lower_part_size':
-				lowersize = float(value.strip())
+				if value.split(',')==['\n']:
+					lowersize = float(0.0)
+				else:
+					lowersize = float(value.strip())
 			if key == 'upper_part_size':
-				uppersize = float(value.strip())
+				if value.split(',')==['\n']:
+					uppersize = float(0.0)
+				else:
+					uppersize = float(value.strip())
 			if key == 'mass_trans_coeff':
-				kgwt = float(value.strip())
+				if value.split(',')==['\n']:
+					kgwt = float(0.0)
+				else:
+					kgwt = float(value.strip())
 			if key == 'eff_abs_wall_massC':
-				Cw = float(value.strip())
+				if value.split(',')==['\n']:
+					Cw = float(0.0)
+				else:
+					Cw = float(value.strip())
 			if key == 'Temperature':
-				TEMP = float(value.strip())
+				if value.split(',')==['\n']:
+					print('Error: no air temperature detected in model inputs file')
+					sys.exit()
+				else:
+					TEMP = float(value.strip())
 			if key == 'PInit':
-				PInit = float(value.strip())
+				if value.split(',')==['\n']:
+					print('Error: no air pressure detected in model inputs file')
+					sys.exit()
+				else:
+					PInit = float(value.strip())
 			if key == 'RH':
-				RH = float(value.strip())
+				if value.split(',')==['\n']:
+					print('Error: no relative humidity detected in model inputs file')
+					sys.exit()
+				else:
+					RH = float(value.strip())
 			if key == 'lat':
-				lat = float(value.strip())
+				if value.split(',')==['\n']:
+					lat = float(0.0)
+				else:
+					lat = float(value.strip())
 			if key == 'lon':
-				lon = float(value.strip())
-			if key == 'daytime_start':
-				dt_start = float(value.strip())
-			if key == 'ChamSA':
-				ChamSA = float(value.strip())
-			if key == 'nucv1':
-				nucv1 = float(value.strip())
+				if value.split(',')==['\n']:
+					lon = float(0.0)
+				else:
[…]
