### 2019-11-21 #1879 @jedbrown
@josephhardinee @dhhagan @andreas-h :wave: Would you be available and interested in reviewing this submission for JOSS?
