### 2019-11-21 #1879 @andreas-h
Yes can do that if it is acceptable that it will take me until the end of the year to properly do it.
