### 2019-11-28 #1918 @whedon
[REVIEW]: PyCHAM: CHemistry with Aerosol Microphysics in Python
**Submitting author:** @simonom (<a href="http://orcid.org/0000-0001-7197-0651">Simon O&#39;Meara</a>)
**Repository:** <a href="https://github.com/simonom/PyCHAM" target ="_blank">https://github.com/simonom/PyCHAM</a>
**Version:** v0.2.4
