### 2019-11-28 #1918 @jedbrown
@dhhagan @andreas-h :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [PyCHAM repository](https://github.com/simonom/PyCHAM)). I'll be watching this thread if you have any questions.
