### 2019-12-04 simonom/PyCHAM#2 @dhhagan
Add community guidelines
There do not appear to be any community guidelines - see [here][1] for an example.

This is in response to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1918).

[1]: https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors
