### 2019-12-04 simonom/PyCHAM#3 @dhhagan
More clearly state how this model is unique
Please add a short clarification on how this model differs from others? You mention that "PyCHAM is novel in its ease of accessibility and utility." Is this because it is the only open-sourced version? The only GUI-driven version? A quick sentence to clear this up would be appreciated!

This is in response to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1918).
