### 2019-12-04 simonom/PyCHAM#4 @dhhagan
Update version
The [JOSS submission][1] is for v0.0.5, but the current GitHub version is 0.2.2 - this should be reconciled on the JOSS submission.

[1]: https://github.com/openjournals/joss-reviews/issues/1918
