### 2019-12-04 simonom/PyCHAM#6 @dhhagan
Improve installation instructions
At this moment, the installation instructions are a bit convoluted. It would be great if you could:

  1. provide links to the non-python packages that are required for installation
  2. include all necessary instructions on this page so that users don't have to navigate multiple other sites
  3. add instructions for installing Sundials and openbabel for different platforms. It looks like sundials is available on [homebrew][1] for macOS. The same looks to be true for [open-babel][2].

I will go ahead and add these instructions and send a PR as part of the review process for [JOSS][3].

[1]: https://formulae.brew.sh/formula/sundials
[2]: https://formulae.brew.sh/formula/open-babel
[3]: https://github.com/openjournals/joss-reviews/issues/1918
