### 2019-12-04 simonom/PyCHAM#7 @dhhagan
assimulo is not listed as a required package in the setup file
https://github.com/openjournals/joss-reviews/issues/1918
