### 2019-12-04 #1918 @dhhagan
Hi @simonom - do you know if the newer versions of SUNDIALS will work? The current version is 5.0, but some of the dependencies list only v2.x - there are also dependency issues associated with python version as well that could possibly cause issues?
