### 2019-12-05 simonom/PyCHAM#2 @simonom
CONTRIBUTING.md file added to master
