+++ paper.md
@@ -40,7 +40,7 @@ bibliography: paper.bib
 
 # Summary
 
-PyCHAM (CHemistry with Aerosol Microphysics in Python) is an open-access O-D box model for environmental chamber studies.  Environmental chambers provide a means for atmospheric scientists to interrogate aerosol processes [@Finlayson:2000; @Schwantes:2017; @Hidy:2019].  Allying PyCHAM with chamber measurements allows quantification of unknown parameters, such as branching ratios for oxidation schemes [@Chen:2005].  Although several box models have already been published [@Pierce:2008; @Roldin:2014; @Sunol:2018; @Roldin:2019], PyCHAM is novel in its ease of accessibility and utility.  The intention therefore, is that it will be readily employed by research groups undertaking environmental chamber measurements.
+PyCHAM (CHemistry with Aerosol Microphysics in Python) is an open-access O-D box model for environmental chamber studies.  Environmental chambers provide a means for atmospheric scientists to interrogate aerosol processes [@Finlayson:2000; @Schwantes:2017; @Hidy:2019].  Allying PyCHAM with chamber measurements allows quantification of unknown parameters, such as branching ratios for oxidation schemes [@Chen:2005].  Although several box models have already been published [@Pierce:2008; @Roldin:2014; @Sunol:2018; @Roldin:2019], PyCHAM is novel in its ease of accessibility and utility.  To the best of our knowledge, PyBOX [@Topping:2018], which formed the basis of PyCHAM, and AtCHEM [@AtCHEM:2020], are the only other box models available online or with a graphical user interface.  However, PyCHAM provides a considerable upgrade to the functionality of both prior models, as AtCHEM only simulates gas-phase chemistry and PyBOX has limited particle effects, whilst PyCHAM models all particle microphysics (detailed below) and wall effects in addition to the gas phase.  The intention therefore, is that it will be readily employed by research groups undertaking environmental chamber measurements.
 
 With air quality and climate models increasingly important to guiding sustainable societies, the accuracy of simulations must suffice [@Tong:2019].  However, research shows that the simulated aerosol effects in these models provides a relatively high amount of uncertainty [@Johnson:2018].  The combination of box models like PyCHAM with environmental chamber measurements to better constrain aerosol processes is therefore necessary to ultimately improve societal sustainability.
 
