### 2019-12-05 simonom/PyCHAM#3 @simonom
Clarification provided in first paragraph of paper.md
