+++ paper.bib
@@ -99,6 +99,14 @@
   urldate = {2018-12-10}
 }
 
+@misc{@AtCHEM:2020,
+	author = {K. Boro{\'n}ska and P. Jimack and M. Pilling and A. Rickard and M. Moreno and D. Waller and J. Young},
+  	title = {AtChem on-line project},
+  	year = 2019,
+  	url = {https://atchem.leeds.ac.uk/webapp/},
+  	urldate = {2019-05-12}
+}
+
 @book{Jacobson:2005,
     author    = {M.Z. Jacobson},
     title     = {Fundamentals of Atmospheric Modeling},
@@ -146,13 +154,4 @@
 	volume = {506},
 	pages = {476--479},
 	doi={10.1038/nature13032},
-}
-
-@misc{@AtCHEM:2020,
-	author = {K. Boro{\'n}ska and P. Jimack and M. Pilling and A. Rickard and M. Moreno and D. Waller and J. Young},
-  	title = {AtChem on-line project},
-  	year = 2020,
-  	url = {https://atchem.leeds.ac.uk/webapp/},
-  	urldate = {2019-05-12}
-}
 }
