+++ paper.bib
@@ -99,7 +99,7 @@
   urldate = {2018-12-10}
 }
 
-@misc{@AtCHEM:2020,
+@misc{AtCHEM:2020,
 	author = {K. Boro{\'n}ska and P. Jimack and M. Pilling and A. Rickard and M. Moreno and D. Waller and J. Young},
   	title = {AtChem on-line project},
   	year = 2019,
