### 2019-12-05 a2b49f9
Install instructions expanded
Links given to the Sundials, Assimulo and open babel download pages
