### 2019-12-05 #1918 @simonom
It's a good question @dhhagan.  I'm looking into this now, whilst updating the installation instructions as you recommended.
