### 2019-12-07 simonom/PyCHAM#6 @simonom
Hi @dhhagan , I have looked into this, as well as the Assimulo dependencies you mentioned (with regard to Sundials and python version).  The Assimulo package represents a bottleneck it seems in terms of install options.  Trying to install Assimulo from source fails in my experience (due to missing files in Assimulo), but works smoothly using the conda installation described in the PyCHAM README.  I suggest that the 'pip Install' section be removed from the README, leaving only the 'conda Install' section.  Would this satisfy the review?
