### 2019-12-08 simonom/PyCHAM#6 @dhhagan
Yup - that works. I would still throw it in as a requirement, as I was able to install pyCHAM directly and it didn't throw an error until runtime.
