### 2019-12-11 simonom/PyCHAM#6 @simonom
Actually, I've found a way to install from source.  So I'll update the instructions today.
