+++ README.md
@@ -4,7 +4,16 @@ Welcome to the PyCHAM.  Funding has been provided from the [EUROCHAMP-2020 resea
 
 PyCHAM is an open-access computer code (written in Python) for simulating aerosol chambers.  It is supplied under the GNU General Public License v3.0.
 
-The programme is controlled via a graphical user interface (GUI).  There are two options for installing, via conda and via pip.  Experience indicates that the conda install is more straightforward than the pip, therefore we recommend this.
+
+# Table of Contents
+1. [Installation](#Installation)
+2. [Running](#Running)
+3. [Testing](#Testing)
+4. [Inputs](#Inputs)
+
+## Installation
+
+There are two options for installing, via conda and from source.  Experience indicates that the conda install is more straightforward than the pip, therefore we recommend this.
 
 
 ## conda Install
@@ -35,7 +44,7 @@ The programme is controlled via a graphical user interface (GUI).  There are two
 
 6. Install complete
 
-## Once Installed
+## Running
 
 1. For model inputs, ensure you have: a .txt file chemical reaction scheme, a .xml file for converting species names to SMILE strings and a .txt file stating values of model variables (e.g. temperature) - see details about these three files below
 
@@ -47,13 +56,15 @@ The programme is controlled via a graphical user interface (GUI).  There are two
 
 5. The 'plot results' button produces (and saves in the output folder) two figures, the first called contours.png that shows the particle number distribution, SOA mass and particle number concentration against time, and the second called gas_pbb.png that shows the gas-phase concentrations of initial components with time  
 
-## Unit Testing, Integration Testing and Example Run
+## Testing
 
 Unit tests for PyCHAM modules can be found in the PyCHAM/Unit_Testing folder in the Github repository.  To use, cd to this folder and use python test_module.py with module replaced by the name of the module to be tested.
 
 Integration testing can be completed using the '.travis.yml' file at the website: travis-ci.org.
 
 The example run output is saved in output/Example_Run/Example_output in the Github repository.  To reproduce this (e.g. for testing), use the inputs/Example_Run.txt for the chemical scheme, inputs/Example_Run_xml.xml for the xml file and inputs/Example_Run_inputs.txt for the model inputs.  Note that the example output was produced using version 0.2.1 of PyCHAM.
+
+## Inputs
 
 ## Chemical Scheme .txt file
 
