### 2019-12-11 33cd95b
Major update to installation instructions
