### 2019-12-11 simonom/PyCHAM#7 @simonom
No, this is because when installing from source, sundials, Assimulo and openbabel require separate installation.  Instructions on how to do the open source installation have now been updated in the readme.
