### 2019-12-20 df73d1c
minor edit to install from source instructions
