### 2020-01-13 #1918 @simonom
Dear @dhhagan and @andreas-h,
Are you satisfied with the provided responses to issues?
