### 2020-03-03 #1918 @dhhagan
Hey @jedbrown Sorry about this - I was trying to find some guidelines for python projects and any/all requirements for JOSS but couldn't seem to find any. Currently, this project likely only works if installed via conda and not via pip due to some extremely challenging/frustrating installation requirements/steps for the required dependencies. Is a conda-only installation acceptable? Are there any more detailed guidelines that are language-specific from JOSS?
