### 2020-03-03 #1918 @jedbrown
@dhhagan The important thing is that the installation instructions work. IMO, this should include some way for a potential contributor to install. JOSS doesn't currently have a "minimum standards" for Python packaging, but it should strive to follow community best practices. The vast majority of Python packages in JOSS will have a working pip install. If there are really good reasons why it is not tractable here and reviewers agree, I wouldn't block acceptance, but there should be something about how a contributor would install. Bad installs tax the community so authors should make an effort to clean up their installs to the extent feasible.
