### 2020-03-03 #1918 @jedbrown
@samjsilva91 :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [PyCHAM repository](https://github.com/simonom/PyCHAM)). I'll be watching this thread if you have any questions.
