### 2020-03-17 #1918 @samjsilva91
Hello,

I apologize for the delay on getting to this review. COVID-19 here in Washington has been a bit hectic!
As I am going through the review process, I can’t actually get the checklist options to work. As in, when I click a check box, it doesn’t actually register a check.
Is there an issue on my end somewhere?

Best,
Sam
