### 2020-03-18 feec710
More detailed added to installation instructions
