### 2020-03-18 149beb7
changed Res_file_name name
So that it doesn't clash with existing file name (and therefore throw an error)
