### 2020-03-18 faf0ac7
Total_model_time consistent with example output
