### 2020-03-19 #1918 @samjsilva91
In my opinion, the basic requirements for each of the check boxes has been met for this software, particularly considering the extensive work done prior to my involvement by @dhhagan

The code installs and runs on my machine, and modifications made to the chemical mechanism lead to results broadly consistent with expected behavior.
