### 2020-03-20 9ba79bf
Improved outputs
Improved names and contents of files from output to be more informative
