### 2020-03-20 c3947e9
Improved output
Improved names and contents of files from output to be more informative
