+++ PyCHAM/saving.py
@@ -5,16 +5,20 @@ import sys
 import numpy as np
 
 def saving(filename, y_mat, t_out, Nresult, x2, numsb, y_mw, num_speci, 
-			savefolder, rbou, Cfactor, MV, testf):
+			savefolder, rbou, Cfactor, MV, testf, spec_namelist):
 			
 	#  -------------------------------
 	# inputs:
 	
 	# filename - name of inputs file name
 	# y_mat - species (columns) concentrations with time (rows) (molecules/cc (air))
-	# Cfactor - conversion factor to change gas-phase concentrations from molecules/cc 
-	# (air) into ppb
+	# x2 - particle radius per size bin (columns), per time step (row) (um) including 
+	#		water
+	# Cfactor - conversion factor to multiply gas-phase concentrations in ppb to get 
+	# molecules/cc (air)
+	# MV - molar volume (cm3/mol) (set in pp_intro)
 	# testf - flag to show whether in normal mode (0) or test mode (1)
+	# spec_namelist - names of species given by the equation file
 	# -------------------------------
 	
 	if testf==1:
@@ -29,32 +33,59 @@ def saving(filename, y_mat, t_out, Nresult, x2, numsb, y_mw, num_speci,
 	# create folder to store results
 	os.makedirs(output_by_sim)
 	
-	
-	# constants
-	
-	const = np.zeros((3, num_speci))
-	const[0,0] = numsb
-	const[0,1] = num_speci
-	const[1, :] = y_mw[:, 0]
-	const[2, :] = MV[:, 0]
-	np.savetxt(os.path.join(output_by_sim,'constants'),const,delimiter=',',
-			header='constants')
+	# dictionary containing constants for model and components
+	const = {}
+	const["number_of_size_bins"] = numsb
+	const["number_of_components"] = num_speci
+	const["molecular_weights_g/mol_corresponding_to_component_names"] = (np.squeeze(y_mw[:, 0]).tolist())
+	const["molecular_volumes_cm3/mol"] = (MV[:, 0].tolist())
+	const["component_names"] = spec_namelist
+	const["factor_for_multiplying_ppb_to_get_molec/cm3"] = Cfactor
+	with open(os.path.join(output_by_sim,'model_and_component_constants'),'w') as f:
+		for key in const.keys():
+			f.write("%s,%s\n"%(key, const[key]))
 	
 	# convert gas-phase concentrations from molecules/cc (air) into ppb
 	# leaving any particle-phase concentrations as molecules/cc (air)
 	y_mat[:, 0:num_speci] = y_mat[:, 0:num_speci]/Cfactor
+	
+	# prepare header for concentrations file
+	y_header = str('')
+	x2_header = str('') # prepare header for files relating to size bins
+	for i in range(numsb+1):
+
+		if i == 0:
+			end = '_g'
+		if i > 0 and i<numsb:
+			end = str('_p'+str(i))
+			if i!=numsb-1:
+				x2_header = str(x2_header+str(i)+', ')
+			else:
+				x2_header = str(x2_header+str(i))
+		if i == numsb:
+			end = '_w'
+		for ii in range(num_speci):
+			if i  == 0 and ii == 0:
+				start = ''
+			else:
+				start = ', '
+			y_header = str(y_header+str(start+spec_namelist[ii])+end)
+			
 	# saving both gas- and particle-phase concentrations of species
-	np.savetxt(os.path.join(output_by_sim,'y'),y_mat,delimiter=',',header='conc_vs_time') 	
[…]
