### 2020-03-23 #1918 @simonom
There is also one point from our internal review, which is changing the wording inside the JOSS paper from 'environmental chamber' to 'aerosol chamber'.  The aerosol chamber is a subset of the environmental chamber.  PyCHAM is designed for aerosol chamber, and whilst it could be employed for any chamber experiment, we do not want to misled readers about its capabilities.  This has been updated in the paper.md file on the master branch of the Github repository.

Finally, the version number with the above modifications is now V0.2.4, whereas the number in the original submission was V0.0.5, so this should be updated in any publication.

Thanks for your reviews @dhhagan and @samjsilva91 .
