### 2020-04-01 b2cac33
users can now use the model variables input file to state times and components that are injected instantaneously and continuously
