+++ PyCHAM/eqn_parser.py
@@ -10,17 +10,31 @@ import pybel
 import formatting
 import xmltodict # for opening and converting xml files to python dictionaries
 import ipdb
+from water_calc import water_calc
 
 # ----------Extraction of eqn info----------
 # Extract the mechanism information
-def extract_mechanism(filename, xmlname, TEMP, PInit, Comp0, testf):
+def extract_mechanism(filename, xmlname, TEMP, PInit, testf, RH, 
+						start_sim_time, lat, lon, act_flux_path, DayOfYear):
 
-	# inputs:
-	# testf - flag for operating in normal mode (0) or testing mode (1)
 	
-	if testf == 1:
-		return(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)
-    	
+	# inputs: ----------------------------------------------------------------------------
+	# testf - flag for operating in normal mode (0) or testing mode (1)
+	# ------------------------------------------------------------------------------------
+	
+	if testf == 1: # for just testing mode
+		return(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)
+    
+    # calculate gas-phase concentrations of M, N2 and O2 (molecules/cc (air))
+	# 1.0e-6 converts from molecules/m3 to molecules/cc
+	# R and Avogadro's constant set the same as in atmosphereFunctions.f90 of AtChem2
+	M_val = (PInit/(8.3144621*TEMP)*6.02214129e+23)*1.0e-6
+	# N2 and O2 given the same multiplication as in atmosphereFunctions.f90 of AtChem2
+	N2_val = M_val*0.7809
+	O2_val = M_val*0.2095
+	# water concentration (C_H2O) (molecules/cc (air))
+	[C_H2O, Psat_water, H2O_mw] = water_calc(TEMP, RH, 6.02214129e+23)
+    
 	# Open the file
 	f_open_eqn = open(filename, mode='r')
 	
@@ -32,21 +46,106 @@ def extract_mechanism(filename, xmlname, TEMP, PInit, Comp0, testf):
 		print('IOError')
 		print('Eqn file not closed')
 		sys.exit()
-
-	# equation list without comments
-	naked_list_eqn = formatting.remove_comments(total_list_eqn)
 	
-	# calculate gas-phase concentrations of M, N2 and O2 (molecules/cc (air))
-	# 1.0e-6 converts from molecules/m3 to molecules/cc
-	M_val = (PInit/(8.3144598*TEMP)*6.0221409e+23)*1.0e-6
-	N2_val = M_val*0.79
-	O2_val = M_val*0.2096
+	naked_list_eqn = [] # empty list for equation reactions
+	RO2_names = [] # empty list for peroxy radicals
+	rrc = [] # empty list for reaction rate coefficients
+	rrc_name = [] # empty list for reaction rate coefficient labels
+	eq_marker = r"%"
+	pr_marker = r"RO2"
+	eqn_flag = 0 # don't collate reaction equations until seen
+	pr_flag = 0 # don't collate peroxy radicals until seen
+	rrc_flag = 0 # don't collate reaction rate coefficients until seen
+	
+	# obtain lists for reaction rate coefficients, peroxy radicals and equation reactions
+	# using chemical scheme input file separators
+	for line in total_list_eqn:
+		
+		line1 = line.strip() # remove bounding white space
+		
+		# --------------------------------------------------------------------------------
+		# generic reaction rate coefficients part
+		# look out for start of coefficients
+		if line1 == str('* Generic Rate Coefficients ;'):
+			rrc_flag = 1 # ready to collate reaction rate coefficients
[…]
