+++ umansysprop/data/__init__.py
@@ -1,0 +1,216 @@
+# vim: set et sw=4 sts=4 fileencoding=utf-8:
+#
+# Copyright (c) 2014 Manchester University.
+# All Rights Reserved.
+
+from __future__ import (
+    unicode_literals,
+    absolute_import,
+    print_function,
+    division,
+    )
+str = type('')
+try:
+    from itertools import izip
+    zip = izip
+except ImportError:
+    pass
+
+
+import re
+
+import pybel
+import pkg_resources as pkg
+
+
+def smarts(s):
+	# uncomment next two lines if using python2, but comment for python3
+#     if not isinstance(s, bytes):
+#         s = s.encode('ascii')
+    try:
+        return pybel.Smarts(s)
+    except IOError as e:
+        # Convert pybel's IOError (?!) into a ValueError
+        raise ValueError(str(e))
+
+
+_parse_re = re.compile(r'^\s*((?P<data>([^#]|\S#)*)(\s+#.*)?|#.*)$')
+def _read_data(
+        filename, key_conv=str, value_conv=float, key_col=0, value_col=1):
+    result = {}
+    cols = None
+    for count, line in enumerate(
+            pkg.resource_stream(__name__, filename), start=1):
+        data = _parse_re.match(line.decode('utf-8')).group('data')
+        if data:
+            data = data.split()
+            try:
+                if cols is None:
+                    cols = len(data)
+                elif len(data) != cols:
+                    raise ValueError(
+                            'Unexpected number of values (expected %d)' % cols)
+                key = key_conv(data[key_col])
+                value = value_conv(data[value_col])
+                if key in result:
+                    raise ValueError(
+                            'Duplicate definition for group %s' % key)
+                result[key] = value
+            except (IndexError, ValueError) as e:
+                e.args += ('on line %d of %s' % (count, filename),)
+                raise
+    return result
+
+
+def _read_smarts(filename):
+    return _read_data(filename, key_conv=int, value_conv=smarts)
+
+
+def _read_matrix(filename, key_conv=str, value_conv=float, symmetric=True):
+    result = {}
+    col_keys = []
+    row_keys = []
+    for count, line in enumerate(
+            pkg.resource_stream(__name__, filename), start = 1):
+        data = _parse_re.match(line.decode('utf-8')).group('data')
+        if data:
+            try:
+                if not col_keys:
+                    col_keys = [key_conv(key) for key in data.split()]
+                else:
+                    row_key, values = data.split(None, 1)
+                    row_key = key_conv(row_key)
+                    if row_key in result:
+                        raise ValueError(
+                                'Duplicate definition for row %s' % key)
+                    values = [value_conv(value) for value in values.split()]
+                    if not len(col_keys) == len(values):
+                        raise ValueError(
+                                'Expected %d values but found %d' % (
+                                    len(col_keys), len(values)))
[…]

[…]
