+++ umansysprop/groups.py
@@ -1,1802 +1,0 @@
-# vim: set et sw=4 sts=4 fileencoding=utf-8:
-#
-# Copyright (c) 2016 David Topping.
-# All Rights Reserved.
-# This file is part of umansysprop.
-#
-# umansysprop is free software: you can redistribute it and/or modify it under
-# the terms of the GNU General Public License as published by the Free Software
-# Foundation, either version 3 of the License, or (at your option) any later
-# version.
-#
-# umansysprop is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-# details.
-#
-# You should have received a copy of the GNU General Public License along with
-# umansysprop.  If not, see <http://www.gnu.org/licenses/>.
-
-
-from __future__ import (
-    unicode_literals,
-    absolute_import,
-    print_function,
-    division,
-    )
-str = type('')
-try:
-    range = xrange
-except NameError:
-    pass
-
-
-from math import ceil
-from itertools import chain, groupby
-from operator import itemgetter
-
-from . import data
-
-def matches(patterns, compound):
-    """
-    Returns a mapping of group identifier to number of matches.
-
-    The *patterns* parameter specifies a dictionary mapping a group number to
-    OpenBabel Smarts objects.
-
-    The *compound* parameter specifies an OpenBabel Molecule object (presumably
-    derived from a SMILES string). The function calculates the number of
-    matches of the Molecule with each Smarts string, returning the result as a
-    mapping.
-
-    :param patterns:
-        The mapping of group numbers to Smarts objects
-
-    :param compound:
-        A Molecule objects to match against the SMARTS strings in *filename*
-
-    :returns:
-        A mapping of group number to match counts
-    """
-    return {
-        group: len(smarts.findall(compound))
-        for group, smarts in patterns.items()
-        }
-
-
-def all_matches(patterns, compounds, bias=True):
-    """
-    Returns a mapping of group identifier to a biased sum of matches.
-
-    The *patterns* parameter specifies a dictionary mapping a group number to
-    OpenBabel Smarts objects.
-
-    The *compounds* parameter specifies a sequence of OpenBabel Molecule
-    objects (presumably derived from SMILES strings) or, if *bias* is True
-    (which it is by default), a mapping of OpenBabel Molecule objects to
-    relative abundances.
-
-    If *bias* is False, the function calculates the number of matches of each
-    Molecule with each SMARTS string, returning a mapping of the sum of all
-    matches with each SMARTS string. If *bias* is True, the function calculates
-    the same sum but each match count will be multiplied by the value
-    associated with the compound in the *compounds* mapping.
-
-    :param patterns:
-        The mapping of group numbers to Smarts objects
-
-    :param compounds:
-        A sequence of OpenBabel Molecule objects
-
-    :param bias:
-        A bool indicating whether or not *compounds* is a mapping (True) and
-        that the value associated with each Molecule should be used as a bias
-        in the resulting sum
-
-    :returns:
-        A mapping of group number to match counts
-    """
-    return {
-        key: sum(count for (group, count) in items)
[…]

[…]
