### 2020-04-01 12f5e76
updated example inputs
