+++ PyCHAM/inputs/apinene_scheme.txt
@@ -1,1177 +1,0 @@
-********************************************************************* ;
-* A citation to the MCM website and the relevant mechanism          * ;
-* construction protocols should be given in any publication using   * ;
-* information obtained from this source, using the following or     * ;
-* comparable wording:                                               * ;
-* The chemical mechanistic information was taken from the Master    * ;
-* Chemical Mechanism, MCM v3.3.1 (ref), via website:                  * ;
-* http://mcm.leeds.ac.uk/MCM.                                       * ;
-* The reference should be: (Jenkin et al., Atmos. Environ., 31, 81, * ;
-* 1997; Saunders et al., Atmos. Chem. Phys., 3, 161, 2003), for     * ;
-* non aromatic schemes; (Jenkin et al., Atmos. Chem. Phys., 3,  * ;
-* 181, 2003; Bloss et al., Atmos. Chem. Phys., 5, 641, 2005), for   * ;
-* aromatic schemes; (Jenkin et al., Atmos. Chem. Phys.,  12, * ;
-* 5275, 2012), for the beta-caryophyllene scheme and (Jenkin et al., ;
-* Atmos. Chem. Phys., 15, 11433, 2015), for the isoprene scheme.  * ;
-********************************************************************* ;
-* MCMv3.3.1 Subset generated for the following species:   ;
-* APINENE ;
-*;
-* Variable definitions.  All species are listed here.;
-*;
-VARIABLE
- C7PAN3 C4PAN6 HCOCO NC72O HCC7CO CH3NO3 H1CO23CHO C106NO3 NC101O2
-CH3OH BIACETO2 CO235C6OOH C109O APINBOH C722OOH O3 APINBOO PAN
-CH3COCH2O H3C2C4CO3H HCOCH2CO3 C89PAN C721CO3H APINBCO APINAO2 C106O
-C4CODIAL HCOCO3H C98O C716OH C721OOH C235C6CO3H NC101O C96O C614O
-C813O2 HSO3 NC71O HCOCH2O2 APINCO CO23C4CHO PINALNO3 CH3O C614O2
-C108NO3 C811O C920CO3 C9PAN2 APINOOB APINOOA C108OOH C413COOOH C716O
-C44OOH NAPINBOOH HO12CO3C4 APINBO2 C614CO C514NO3 CHOC3COPAN C621O
-OH C5PAN9 BIACETOH HONO APINBNO3 NC102OOH C717O2 C312COCO3H C89O
[…]
