+++ PyCHAM/inputs/CH4sim_scheme.txt
@@ -1,74 +1,0 @@
-//Extracted from MCM3.3.1: http://mcm.leeds.ac.uk/MCMv3.3.1/extract.htt, including:
-//full inorganic list, CH4 marked
-//EQUATIONS
-% 5.6D-34*N2*(TEMP/300)@-2.6*O2 : O = O3 ;
-% 6.0D-34*O2*(TEMP/300)@-2.6*O2 : O = O3 ;
-% 8.0D-12*EXP(-2060/TEMP) : O + O3 = ;
-% KMT01 : O + NO = NO2 ;
-% 5.5D-12*EXP(188/TEMP) : O + NO2 = NO ;
-% KMT02 : O + NO2 = NO3 ;
-% 3.2D-11*EXP(67/TEMP)*O2 : O1D = O ;
-% 2.0D-11*EXP(130/TEMP)*N2 : O1D = O ;
-% 1.4D-12*EXP(-1310/TEMP) : NO + O3 = NO2 ;
-% 1.4D-13*EXP(-2470/TEMP) : NO2 + O3 = NO3 ;
-% 3.3D-39*EXP(530/TEMP)*O2 : NO + NO = NO2 + NO2 ;
-% 1.8D-11*EXP(110/TEMP) : NO + NO3 = NO2 + NO2 ;
-% 4.50D-14*EXP(-1260/TEMP) : NO2 + NO3 = NO + NO2 ;
-% KMT03 : NO2 + NO3 = N2O5 ;
-% 2.14D-10*H2O : O1D = OH + OH ;
-% 1.70D-12*EXP(-940/TEMP) : OH + O3 = HO2 ;
-% 7.7D-12*EXP(-2100/TEMP) : OH + H2 = HO2 ;
-% KMT05 : OH + CO = HO2 ;
-% 2.9D-12*EXP(-160/TEMP) : OH + H2O2 = HO2 ;
-% 2.03D-16*(TEMP/300)@4.57*EXP(693/TEMP) : HO2 + O3 = OH ;
-% 4.8D-11*EXP(250/TEMP) : OH + HO2 = ;
-% 2.20D-13*KMT06*EXP(600/TEMP) : HO2 + HO2 = H2O2 ;
-% 1.90D-33*M*KMT06*EXP(980/TEMP) : HO2 + HO2 = H2O2 ;
-% KMT07 : OH + NO = HONO ;
-% KMT08 : OH + NO2 = HNO3 ;
-% 2.0D-11 : OH + NO3 = HO2 + NO2 ;
-% 3.45D-12*EXP(270/TEMP) : HO2 + NO = OH + NO2 ;
-% KMT09 : HO2 + NO2 = HO2NO2 ;
-% 3.2D-13*EXP(690/TEMP)*1.0 : OH + HO2NO2 = NO2 ;
-% 4.0D-12 : HO2 + NO3 = OH + NO2 ;
-% 2.5D-12*EXP(260/TEMP) : OH + HONO = NO2 ;
-% KMT11 : OH + HNO3 = NO3 ;
-% 4.0D-32*EXP(-1000/TEMP)*M : O + SO2 = SO3 ;
-% KMT12 : OH + SO2 = HSO3 ;
-% 1.3D-12*EXP(-330/TEMP)*O2 : HSO3 = HO2 + SO3 ;
-% 6.00D-06 : HNO3 = NA ;
-% 4.00D-04 : N2O5 = NA + NA ;
-% 1.20D-15*H2O : SO3 = SA ;
-% J<1> : O3 = O1D ;
-% J<2> : O3 = O ;
-% J<3> : H2O2 = OH + OH ;
[…]

[…]
