### 2020-04-01 852480b
updated PyCHAM_CI_test.py
