+++ PyCHAM/PyCHAM_CI_test.py
@@ -31,97 +31,419 @@ inputs = open(inname, mode='r')
 in_list = inputs.readlines()
 inputs.close()
 
-if len(in_list) != 36:
-	print('Error: The number of model inputs is incorrect, should be 36, but is ' + str(len(in_list)) )
+if len(in_list) != 55:
+	print('Error: The number of variables in the model variables file is incorrect, should be 55, but is ' + str(len(in_list)) )
 	sys.exit()
 for i in range(len(in_list)):
 	key, value = in_list[i].split('=')
-	key = key.strip() # a string
+	key = key.strip() # a string with bounding white space removed
 	if key == 'Res_file_name':
+		if value.split(',')==['\n']:
+			print('Error: no requested results file name detected in model inputs file, please supply')
+			sys.exit()
 		resfname = str(value.strip())
 	if key == 'Total_model_time':
-		end_sim_time = float(value.strip())
-	if key == 'Time_step':
-		tstep_len = float(value.strip())
-	if key == 'Recording_time_step':
-		save_step = float(value.strip())
-	if key == 'Number_size_bins':
-		num_sb = int(value.strip())
-	if key == 'lower_part_size':
-		lowersize = float(value.strip())
-	if key == 'upper_part_size':
-		uppersize = float(value.strip())
-	if key == 'mass_trans_coeff':
-		kgwt = float(value.strip())
-	if key == 'eff_abs_wall_massC':
-		Cw = float(value.strip())
-	if key == 'Temperature':
-		TEMP = float(value.strip())
-	if key == 'PInit':
-		PInit = float(value.strip())
-	if key == 'RH':
-		RH = float(value.strip())
-	if key == 'lat':
-		lat = float(value.strip())
-	if key == 'lon':
-		lon = float(value.strip())
-	if key == 'daytime_start':
-		dt_start = float(value.strip())
-	if key == 'ChamSA':
-		ChamSA = float(value.strip())
-	if key == 'nucv1':
-		nucv1 = float(value.strip())
-	if key == 'nucv2':
-		nucv2 = float(value.strip())
-	if key == 'nucv3':
-		nucv3 = float(value.strip())
-	if key == 'nuc_comp':
-		nuc_comp = int(value.strip())
-	if key == 'inflectDp':
-		inflectDp = float(value.strip())
-	if key == 'Grad_pre_inflect':
-		pwl_xpre = float(value.strip())
-	if key == 'Grad_post_inflect':
-		pwl_xpro = float(value.strip())
-	if key == 'Kern_at_inflect':
-		inflectk = float(value.strip())
-	if key == 'Rader_flag':
-		Rader = int(value.strip())
-	if key == 'C0':
-		C0 = [float(i) for i in (value.split(','))]
-	if key == 'Comp0': # strip removes white space
-		Comp0 = [str(i).strip() for i in (value.split(','))]
-	if key == 'voli':
-		
 		if value.split(',')==['\n']:
+			print('Error: no total model run time detected in model inputs file, please supply')
+			sys.exit()
+		else:
+			end_sim_time = float(value.strip())
+	if key == 'Time_step':
+		if value.split(',')==['\n']:
+			print('Notice: No model time step detected in model inputs file, defaulting to 60s')
+			tstep_len = float(60.0)
+		else:
+			tstep_len = float(value.strip())
+	if key == 'Recording_time_step':
+		if value.split(',')==['\n']:
+			print('Notice: no recording time step detected in model inputs file, defaulting to 60s')
+			save_step = float(60.0)
+		else:
+			save_step = float(value.strip())
+	if key == 'Number_size_bins':
+		if value.split(',')==['\n']:
[…]
