### 2020-04-01 6090f4d
adapted setup.py fpr Travis
