### 2020-04-02 21fa28a
users specify chemical scheme formatting
