+++ PyCHAM/user_input.py
@@ -37,7 +37,7 @@ def run(source, testf):
 			dydt_trak, space_mode, Ct, Compt, injectt, seed_name, 
 			const_comp, const_infl, Cinfl, act_wi, act_w, seed_mw, 
 			umansysprop_update, core_dens, p_char, e_field, 
-			const_infl_t] = pickle.load(pk)	
+			const_infl_t, chem_scheme_markers] = pickle.load(pk)	
 
 			
 			# convert chamber surface area (m2) to spherical equivalent radius (m)
@@ -70,7 +70,8 @@ def run(source, testf):
 		pwl_xpre, pwl_xpro, inflectk, xmlname, C0, Comp0, Rader, voli, volP, pconc, std, 
 		mean_rad, core_diss, light_stat, light_time, kgwt, 0, dydt_trak, DayOfYear, 
 		space_mode, Ct, Compt, injectt, seed_name, const_comp, const_infl, Cinfl, act_wi, 
-		act_w, seed_mw, umansysprop_update, core_dens, p_char, e_field, const_infl_t)
+		act_w, seed_mw, umansysprop_update, core_dens, p_char, e_field, const_infl_t, 
+		chem_scheme_markers)
 		
 	if source == 1:
 		return(fname, resfname, y_indx_plot, Comp0)
