+++ PyCHAM/Rate_coeffs.py
@@ -10,7 +10,7 @@
 #                                                                                                   # 
 ##################################################################################################### 
 # Minor modified by XSX
-# File Created at 2020-04-02 09:13:35.139216
+# File Created at 2020-04-02 12:12:40.304099
 
 import numpy
 import PhotolysisRates
