+++ PyCHAM_macenv.yml
@@ -72,6 +72,7 @@ dependencies:
   - pytz=2019.3=py_0
   - qt=5.9.7=h468cd18_1
   - readline=8.0=hcfe32e1_0
+  - requests
   - scipy=1.3.1=py37hab3da7d_2
   - setuptools=41.6.0=py37_1
   - sip=4.19.8=py37h0a44026_0
