+++ PyCHAM_linenv.yml
@@ -78,6 +78,7 @@ dependencies:
   - pytz=2019.2=py_0
   - qt=5.9.7=h5867ecd_1
   - readline=8.0=hf8c457e_0
+  - requests
   - scipy=1.3.1=py37h921218d_2
   - setuptools=41.2.0=py37_0
   - sip=4.19.8=py37hf484d3e_1000
