+++ PyCHAM/PyCHAM.py
@@ -480,7 +480,7 @@ class PyCHAM(QWidget):
 					else:
 						print('Error: user has requested cloning of UManSysProp via the model variables input file but connection to the page failed, possibly due to no internet connection (UManSysProp repository site: https://github.com/loftytopping/UManSysProp_public.git)')
 						sys.exit()
-			f key == 'chem_scheme_markers': # formatting for chemical scheme
+			if key == 'chem_scheme_markers': # formatting for chemical scheme
 				if (value.strip()).split(',')==['']:
 					# default to MCM inputs
 					chem_scheme_markers = str['* Reaction definitions. ;', '%', '(.*) End (.*)', '* Generic Rate Coefficients ;', ';', '\*\*\*\*', 'RO2', '*;']
