### 2020-04-02 89389ab
corrected bug in volat_calc.py that prevented seeing umansysprop
