+++ PyCHAM/Rate_coeffs.py
@@ -10,7 +10,7 @@
 #                                                                                                   # 
 ##################################################################################################### 
 # Minor modified by XSX
-# File Created at 2020-04-02 12:12:40.304099
+# File Created at 2020-04-02 15:34:51.144128
 
 import numpy
 import PhotolysisRates
