+++ PyCHAM/inputs/Example_Run_inputs.txt
@@ -1,4 +1,4 @@
-Res_file_name = Example_output2
+Res_file_name = Example_output
 Total_model_time = 14400.0
 Time_step = 60.0
 Recording_time_step = 60.0
