### 2020-04-03 ad86588
corrected spelling errors
