### 2020-04-03 1a8007a
corrected gas-phase chemistry equation
added with multiplying factor i_{s}
