### 2020-04-03 76c0c86
fixed bug in saving and made unit test for res_plot_super.py
