### 2020-04-03 29ee54b
fixed bug that was giving divide by zero error in pp_intro.py
