+++ PyCHAM/Rate_coeffs.py
@@ -10,7 +10,7 @@
 #                                                                                                   # 
 ##################################################################################################### 
 # Minor modified by XSX
-# File Created at 2020-04-03 14:06:32.412282
+# File Created at 2020-04-03 14:54:51.808624
 
 import numpy
 import PhotolysisRates
