### 2020-04-03 2b6b25f
corrected version in README and run time in Example inputs
