### 2020-04-06 a9f424d
changed index inputs to component names inputs
