+++ PyCHAM/PyCHAM.py
@@ -62,7 +62,7 @@ class PyCHAM(QWidget):
 	def on_click1(self):
 		
 		dirpath = os.getcwd() # get current path
-		
+# 		fname = dirpath+'/PyCHAM/inputs/MAC_Chem.txt' # hard-code chemical scheme input
 		fname = self.openFileNameDialog() # ask for location of input chemical scheme file
 		with open(dirpath+'/fname.txt','w') as f:
 			f.write(fname)
@@ -72,7 +72,7 @@ class PyCHAM(QWidget):
 	def on_click2(self):
 		
 		dirpath = os.getcwd() # get current path
-		
+# 		xmlname = dirpath+'/PyCHAM/inputs/Example_Run_xml.xml' # hard-code xml input
 		xmlname = self.openFileNameDialog()
 		with open(dirpath+'/xmlname.txt','w') as f:
 			f.write(xmlname)
@@ -80,7 +80,8 @@ class PyCHAM(QWidget):
 		
 	@pyqtSlot()
 	def on_click3(self):
-		
+		dirpath = os.getcwd() # get current path
+# 		inname = dirpath+'/PyCHAM/inputs/MAC_inputs.txt' # hard-code model variables input
 		inname = self.openFileNameDialog() # name of model variables inputs file
 		
 		# open the file
@@ -90,8 +91,8 @@ class PyCHAM(QWidget):
 		in_list = inputs.readlines()
 		inputs.close()
 		
-		if len(in_list) != 56:
-			print('Error: The number of variables in the model variables file is incorrect, should be 56, but is ' + str(len(in_list)) )
+		if len(in_list) != 57:
+			print('Error: The number of variables in the model variables file is incorrect, should be 57, but is ' + str(len(in_list)) )
 			sys.exit()
 		for i in range(len(in_list)):
 			key, value = in_list[i].split('=')
@@ -215,11 +216,11 @@ class PyCHAM(QWidget):
 					nucv3 = float(0.0)
 				else:
 					nucv3 = float(value.strip())
-			if key == 'nuc_comp': # index of the nucleating component
+			if key == 'nuc_comp': # name of nucleating component
 				if (value.strip()).split(',')==['']:
-					nuc_comp = float(0.0)
+					nuc_comp = [] # empty list
 				else:
-					nuc_comp = int(value.strip())
+					nuc_comp = [str(i).strip() for i in (value.split(','))]
 			if key == 'new_partr': # radius of newly nucleated particles (cm)
 				if value.split(',')==['\n']:
 					new_partr = float(2.0e-7)
@@ -366,11 +367,11 @@ class PyCHAM(QWidget):
 						Cinfl[i, :] = [float(ii.strip()) for ii in ((value.split(';')[i]).split(','))]
 			
 			
-			if key == 'voli':
+			if key == 'vol_Comp':
 				if (value.strip()).split(',')==['']:
-					voli = np.empty(0)
+					vol_Comp = [] # empty list
 				else:
-					voli = [int(i) for i in (value.split(','))]	
+					vol_Comp = [str(i).strip() for i in (value.split(','))]	
 			if key == 'volP':
 				if (value.strip()).split(',')==['']:
 					volP = np.empty(0)
@@ -480,9 +481,15 @@ class PyCHAM(QWidget):
 			if key == 'chem_scheme_markers': # formatting for chemical scheme
 				if (value.strip()).split(',')==['']:
 					# default to MCM inputs
-					chem_scheme_markers = str['* Reaction definitions. ;', '%', '(.*) End (.*)', '* Generic Rate Coefficients ;', ';', '\*\*\*\*', 'RO2', '*;']
[…]
