### 2020-04-06 d07d99d
changed index inputs to component names inputs
