+++ PyCHAM/inputs/MAC_Chem.txt
@@ -1,1192 +1,0 @@
-#RO2_BEGIN
-    NAPINAO2;
-    NAPINBO2;
-    APINAO2;
-    APINBO2;
-    APINCO2;
-    C107O2;
-    C109O2;
-    C96O2;
-    NC101O2;
-    C96CO3;
-    C720O2;
-    PINALO2;
-    C108O2;
-    C89CO3;
-    C920CO3;
-    C920O2;
-    C97O2;
-    C85CO3;
-    C85O2;
-    CH3COCH2O2;
-    CH3CO3;
-    CH3O2;
-    C719O2;
-    NC102O2;
-    C106O2;
-    C717O2;
-    C811CO3;
-    C89O2;
-    C921O2;
-    C98O2;
-    C86O2;
-    CO235C6CO3;
-    CHOC3COCO3;
-    NC71O2;
-    C811O2;
-    CHOC3COO2;
-    H3C25C6CO3;
-    H3C25C6O2;
-    CO235C6O2;
-    C716O2;
-    C810O2;
-    C922O2;
-    C614O2;
-    C511O2;
-    C812O2;
-    C721CO3;
-    C721O2;
-    HCOCH2CO3;
-    BIACETO2;
-    HOCH2CO3;
-    H3C2C4CO3;
-    HMVKAO2;
-    CO23C4CO3;
-    C312COCO3;
-    CHOCOCH2O2;
-    NC72O2;
-    C514O2;
-    HCOCH2O2;
-    C621O2;
-    C813O2;
-    C722O2;
-    CO2H3CO3;
-    NC61CO3;
-    HCOCO3;
-    C516O2;
-    C44O2;
-    H1C23C4CO3;
-    H1C23C4O2 ;;
-#RO2_END
-#GENERIC_RATE_COEFFICIENTS_BEGIN
-    KRO2NO = 2.7e-12*EXP(360/TEMP);
-    KRO2HO2 = 2.91e-13*EXP(1300/TEMP);
-    KAPHO2 = 5.2e-13*EXP(980/TEMP);
-    KAPNO = 7.5e-12*EXP(290/TEMP);
-    KRO2NO3 = 2.3e-12;
-    KNO3AL = 1.4e-12*EXP(-1860/TEMP);
-    KDEC = 1.00e+06;
-    KROPRIM = 2.50e-14*EXP(-300/TEMP);
-    KROSEC = 2.50e-14*EXP(-300/TEMP);
-    KCH3O2 = 1.03e-13*EXP(365/TEMP);
-    K298CH3O2 = 3.5e-13;
-    K14ISOM1 = 3.00e7*EXP(-5300/TEMP);
-    KD0 = 1.10e-05*M*EXP(-10100/TEMP);
-    KDI = 1.90e17*EXP(-14100/TEMP);
-    KRD = KD0/KDI;
-    FCD = 0.30;
-    NCD = 0.75-1.27*(LOG10(FCD));
-    FD = 10**(LOG10(FCD)/(1+(LOG10(KRD)/NCD)**2));
-    KBPAN = (KD0*KDI)*FD/(KD0+KDI);
-    KC0 = 3.28e-28*M*(TEMP/300)**-6.87;
-    KCI = 1.125e-11*(TEMP/300)**-1.105;
-    KRC = KC0/KCI;
-    FCC = 0.30;
-    NC = 0.75-1.27*(LOG10(FCC));
-    FC = 10**(LOG10(FCC)/(1+(LOG10(KRC)/NC)**2));
-    KFPAN = (KC0*KCI)*FC/(KC0+KCI);
[…]
