+++ PyCHAM/inputs/MAC_inputs.txt
@@ -1,57 +1,0 @@
-Res_file_name = MAC_20120215_16sb
-Total_model_time = 14400.0
-Time_step = 60.0
-Recording_time_step = 60.0
-Number_size_bins = 16
-lower_part_size = 0.0
-upper_part_size = 8.0e-2
-space_mode = lin
-mass_trans_coeff = 0.00575
-eff_abs_wall_massC = 1e0
-Temperature = 294.15
-PInit = 103556.72
-RH = 0.65
-lat =
-lon =
-DayOfYear =
-daytime_start =
-act_flux_path =
-ChamSA = 42 
-nucv1 = 31000.0
-nucv2 = -55.0
-nucv3 = 180.0
-nuc_comp = nuc
-new_partr = 5.270095473654397e-07
-inflectDp = 2.0e-7
-Grad_pre_inflect = 0.0
-Grad_post_inflect = -7.0e-5
-Rate_at_inflect = 9.5e-5
-part_charge_num =
-elec_field =
-McMurry_flag = 0
-C0 = 25.0, 30.0, 0.05, 1.8
-Comp0 = O3,APINENE,NO,NO2
-injectt = 
-Compt = 
-Ct = 
-const_comp =
-const_infl =
-const_infl_t =
-Cinfl =
-vol_Comp = ELVOC_o3, ELVOC_oh, ELVOC_pinal, ELVOC_c107ooh, ELVOC_c109co, ELVOC_norpinal, ELVOC_pinonic, ELVOC_c96ooh, nuc, SVOC_o3, SVOC_oh, SVOC_pinal, SVOC_c107ooh, SVOC_c109co, SVOC_norpinal, SVOC_pinonic, SVOC_c96ooh
-volP = 1.0e-30, 1.0e-30, 1.0e-30, 1.0e-30, 1.0e-30, 1.0e-30, 1.0e-30, 1.0e-30, 1.0e-30, 6e-7, 6e-7, 6e-7, 6e-7, 6e-7, 6e-7, 6e-7, 6e-7
-act_wi =
-act_w =
-pconc = 0.0
-seed_name =
-seed_mw =
-seed_dens =
-mean_rad =
-std =
-core_diss =
-light_time = 
-light_stat =
-tracked_comp = 
-umansysprop_update =
-chem_scheme_markers = #EQUATIONS_BEGIN,     %, #EQUATIONS_END, #GENERIC_RATE_COEFFICIENTS_BEGIN, ;, #GENERIC_RATE_COEFFICIENTS_END, #RO2_BEGIN, ;, #RO2_END
-int_tol = 1.0e-3, 1.0e-4 
