### 2020-04-08 ca983d4
added Sundials and corrected abbreviations
