+++ paper.bib
@@ -27,7 +27,7 @@
 @article{Chen:2005,    
 	author = {J. Chen and R.J. Griffin},
 	title = {Modeling secondary organic aerosol formation from oxidation of $\alpha$-pinene, $\beta$-pinene and $d$-limonene},
-	journal = {Atmos. Env.},
+	journal = {Atmospheric Environment},
 	year = {2005},
 	volume = {39},
 	pages = {7731-7744},
@@ -37,7 +37,7 @@
 @article{Pierce:2008,    
 	author = {J. Pierce and G. Englehart and L. Hildebrandt and E. Weitkamp and E. Parthak and N. Donahue},
 	title = {Constraining Particle Evolution from Wall Losses, Coagulation and Condensation-Evaporation in Smog Chamber Experiments: Optimal Based Size Distribution Measurements},
-	journal = {Aerosol Sci. Tech.},
+	journal = {Aerosol Science and Technology},
 	year = {2008},
 	volume = {42},
 	doi = {10.1080/02786820802389251},
@@ -46,7 +46,7 @@
 @article{Roldin:2014,    
 	author = {P. Roldin and A.C. Ericsson and E.Z. Nordin and E. Hermansson and D. Mogensen and A. Rusanen and M. Boy and E. Swietlicki and B. Svenningsson and A. Zelenyuk and J. Pagels},
 	title = {Modelling non-equilibrium secondary organic aerosol formation and evaporation with the aerosol dynamics, gas- and particle-phase chemistry kinetic multilayer model ADCHAM},
-	journal = {Atmos. Chem. Phys.},
+	journal = {Atmospheric Chemistry and Physics},
 	year = {2014},
 	volume = {14},
 	doi = {10.5194/acp-14-7953-2014},
@@ -55,7 +55,7 @@
 @article{Sunol:2018,    
 	author = {A.M. Sunol and S.M. Charan and J.H. Seinfeld},
 	title = {Computational simulation of the dynamics of secondary organic aerosol formation in an environmental chamber},
-	journal = {Aerosol Sci. Tech.},
+	journal = {Aerosol Science and Technology},
 	year = {2018},
 	pages = {470 - 482},
 	volume = {52},
@@ -65,7 +65,7 @@
 @article{Roldin:2019,    
 	author = {P. Roldin and M. Ehn and T. Kurt{\'e}n and T. Olenius and M.P. Rissanen and N. Sarnela and J. Elm and P. Rantala and L. Hao and N. Hyttinen and and L. Heikkinen and D.R. Worsnop and L. Pichelstorfer and C. Xavier and P. Clusius and E. {\"O}str{\"o}m and T. Pet{\"a}j{\"a} and M. Kulmala and H. Vehkam{\"a}ki and A. Virtanen and I. Riipinen and M. Boy},
 	title = {The role of highly oxygenated organic molecules in the Boreal aerosol-cloud-climate system},
-	journal = {Nat. Commun.},
+	journal = {Nature Communications},
 	year = {2019},
 	volume = {10},
 	doi = {10.1038/s41467-019-12338-8},
@@ -74,7 +74,7 @@
 @article{Tong:2019,    
 	author = {D. Tong and G. Geng and K. Jiang and J. Cheng and Y. Zheng and C. Hong and L. Yan and Y. Zhang and X. Chen and Y. Bo and Y. Lei and Q. Zhang and K. He},
 	title = {Energy and emission pathways towards PM$_{2.5}$ air quality attainment in the Beijing-Tianjin-Hebei region by 2030},
-	journal = {Sci. Total Environ.},
+	journal = {Science of the Total Environment},
 	year = {2019},
 	volume = {692},
[…]
