### 2020-04-08 ec697ac
in response to Jed Brown's comments
moved Eurochamp reference to acknowledgements and cited Sundials
