### 2020-04-08 #1918 @simonom
Thanks Jed, have included the improvements you suggested (Sundials citation, non-abbreviated journal titles and moving EUROCHAMP reference to acknowledgements).
