### 2020-04-08 1656836
users can set their own files for absorption cross section and quantum yield
