### 2020-04-15 e8ed6a5
updated version number (v0.2.4)
Now consistent with JOSS paper
