### 2020-04-15 #1918 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
